Update logger connector
